backend/requirements.txt
# Core Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.2
pydantic-settings==2.6.1

# Portia AI with Gemini support (OFFICIAL)
portia-sdk-python[google]==1.2.0

# Database & ORM
prisma==0.15.0
asyncpg==0.29.0

# Authentication (Clerk integration)
clerk-sdk-python==2.8.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# HTTP & External APIs
httpx==0.28.0
requests==2.32.3

# Async & Background Tasks  
celery[redis]==5.4.0
redis==5.2.0

# Environment & Config
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.4.0
rich==13.9.4

# Email & Communication
google-api-python-client==2.153.0
slack-sdk==3.33.3

# Security
bcrypt==4.2.1
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0